Trait grabbag::iter::pad_tail_to::PadTailToIterator

pub trait PadTailToIterator<E>: Iterator<Item=E> + Sized {
    fn pad_tail_to<F: FnMut(usize) -> E>(self, n: usize, filler: F) -> PadTailTo<Self, F> { ... }
}

Provided Methods

fn pad_tail_to<F: FnMut(usize) -> E>(self, n: usize, filler: F) -> PadTailTo<Self, F>

Creates an iterator that ensures there are at least n elements in a sequence. If the input iterator is too short, the difference is made up with a filler value.