Trait grabbag::iter::zip_longest::ZipLongestIterator

pub trait ZipLongestIterator: Iterator + Sized {
    fn zip_longest<RightIt>(self, right: RightIt) -> ZipLongest<Self, RightIt> { ... }
}

Provided Methods

fn zip_longest<RightIt>(self, right: RightIt) -> ZipLongest<Self, RightIt>

Creates an iterator which yields elements from both input iterators in lockstep. If one iterator ends before the other, the elements from that iterator will be replaced with None.