Trait grabbag::iter::accumulate::AccumulateIterator

pub trait AccumulateIterator<E>: Iterator<Item=E> + Sized {
    fn accumulate<F: FnMut(E, E) -> E>(self, f: F) -> Accumulate<Self, E, F> { ... }
}

Provided Methods

fn accumulate<F: FnMut(E, E) -> E>(self, f: F) -> Accumulate<Self, E, F>

Creates an iterator that scans from left to right over the input sequence, returning the accumulated result of calling f on the entire sequence up to that point.