Trait grabbag::iter::zip_longest::ZipLongestIterator

pub trait ZipLongestIterator: Iterator + Sized {
    fn zip_longest<RightIt>(self, right: RightIt) -> ZipLongest<Self, RightIt> { ... }
}

(..., am), (..., bm, bm+1, ..., bn) → ( ..., (Some(am), Some(bm)), (None, Some(bm+1)), ..., (None, Some(bn)) )

Provided Methods

fn zip_longest<RightIt>(self, right: RightIt) -> ZipLongest<Self, RightIt>

Creates an iterator which yields elements from both input iterators in lockstep. If one iterator ends before the other, the elements from that iterator will be replaced with None.